Create memory manipulation instructions
Create and test instructions for manipulating memory. Also fix an issue with
memory cells with a single child.

I've decided that the top-level memory cell in a program will be inaccessible
from within the program, and will just be there for algorithmic convenience. If
a program tries to move to the top-level cell, it will exit.