Update parser
Implement string literal concatenation syntax using `&`. Allow any literal to be
used as the initial memory state, not just memory block literals. Refactor
parsing code to be nicer and easier to follow. Use move semantics to make it
clear that a Program's input stream should not be read from after it's
constructed. Improve tests for the parser.